Cleaned up projects to be more organized & fixed issue with 404 project
