+++ frontend/src/pages/projects/screens/projects/ProjectView.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import ApplicationsPage from '../../../ApplicationsPage';
-import EmptyProjects from '../../EmptyProjects';
+import EmptyProjects from './EmptyProjects';
 import useUserProjects from './useUserProjects';
 import ProjectListView from './ProjectListView';
